requirements.txt
# Binary QA Framework Requirements
# =================================

# Core ML Frameworks
# -----------------
tensorflow==2.15.0  # For LSTM models
torch==2.1.0        # For Transformer models and BoTorch

# Data Processing
# --------------
numpy==1.26.4       # Numerical computing
pandas==2.2.1       # Data manipulation
scipy==1.11.3       # Scientific computing

# Visualization
# ------------
matplotlib==3.8.1   # Basic plotting
seaborn==0.13.0     # Statistical visualizations

# Bayesian Optimization
# -------------------
ax-platform==0.2.8  # Ax platform for optimization
botorch==0.6.2      # Bayesian optimization with PyTorch

# Utilities
# --------
tqdm==4.65.0        # Progress bars
pyyaml==6.0.1       # Configuration handling
scikit-learn==1.3.1 # Metrics and evaluation utilities

# Experiment Tracking
# -----------------
neptune-client==1.13.0  # Experiment logging and tracking

# Development Environment
# ---------------------
jupyter>=1.0.0      # Jupyter notebooks